desktop: Use internal wayland-modules
