+++ src/backend/renderer/mod.rs
@@ -37,7 +37,6 @@ use crate::backend::egl::{
 #[cfg(feature = "renderer_multi")]
 pub mod multigpu;
 
-#[cfg(feature = "wayland_frontend")]
 pub mod utils;
 
 pub mod element;
